Change no-section handling
When you are doing the manual filter, or _registering_ the callback, it will still pop up the error message and stop the filtering process (unregistering as needed).

However, if it's during the callback that it hits an unknown type, it will just render the error message to the filtered browser instead.  This will be more graceful (so if you save a non-filtered type while the callback is active, the callback will stay active)